requirements.txt
Flask==2.2.2  
flask-socketio 
Jinja2==3.1.2  
Werkzeug==2.2.2
selenium
pillow
openai
markdown2
latex2mathml
psutil

